backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
groq
httpx
python-dotenv
bcrypt
python-jose[cryptography]
passlib
python-multipart
numpy
